보강 #5(1/2): 금액·수량 BigDecimal — Flink 모델·파서·싱크·Binance 3파일, 비율은 Float64 캐스팅, Decimal 컷오버 런북
- CryptoTradeEvent/BinanceTrade 의 price·volume·qty 를 BigDecimal 로, amount·quote_qty 는 price×volume 계산(MySQL DECIMAL(20,4) 의 먼지 0 문제 제거)
- MarketAlertDetector 는 double 유지(비율 판정 + ValueState<double[]> 세이브포인트 호환) — 테스트 9/9 그대로 통과가 증거
- dbt 7모델: Decimal 나눗셈은 분모 0 에서 예외를 던져 모델이 죽는다(if 가드도 무용) → 비율 전부 toFloat64 + nullIf
- 런북: Float64→Decimal 은 toString 경유 필수(CAST 는 버림 — KRW-PEPE 42.86 KRW 오차로 적발)
- 테스트 19/19 (먼지 체결 회귀 테스트 추가)

+++ flink/src/main/java/com/cdc/pipeline/function/MarketAlertDetector.java
@@ -92,7 +92,10 @@ public class MarketAlertDetector extends KeyedProcessFunction<String, CryptoTrad
             return;
         }
         long m = Math.floorDiv(e.getUpbitTimestamp(), MINUTE_MS);
-        double p = e.getTradePrice();
+        // 2026-09-20 (docs/34 #5): 이벤트는 BigDecimal 이지만 **여기서만 double 로 받는다**.
+        // 이유 ① 판정이 비율 비교(±50/100/200%)라 double 로도 결과가 같다 ② 상태(ValueState<double[]> 1,500분 링)의 타입을 바꾸면
+        // 세이브포인트 복원이 깨진다 — 규칙 동등성(docs/22 §4-5, 116/116)을 유지하려면 상태 타입을 건드리면 안 된다.
+        double p = e.getTradePrice().doubleValue();
         if (p <= 0) return;
 
         double[] r = ring.value();
