+++ flink/src/main/java/com/cdc/pipeline/model/CryptoTradeEvent.java
@@ -1,6 +1,7 @@
 package com.cdc.pipeline.model;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * 암호화폐 체결 이벤트 모델
@@ -14,9 +15,12 @@ public class CryptoTradeEvent implements Serializable {
     private String op;              // CDC 오퍼레이션: r(snapshot), c(insert), u(update), d(delete)
     private long tradeId;           // MySQL AUTO_INCREMENT PK
     private String market;          // 마켓 코드 (KRW-BTC, KRW-ETH 등)
-    private double tradePrice;      // 체결 가격 (KRW)
-    private double tradeVolume;     // 체결 수량
-    private double tradeAmount;     // 체결 금액 (price × volume)
+    // 2026-09-20 (docs/34 #5): 금액·수량은 BigDecimal. 원천이 MySQL DECIMAL(20,8)·Debezium decimal.handling.mode=string 이라
+    // 정밀도는 문자열로 도착하는데 double 로 받아 버리고 있었다(저장 층 타입은 하류에 주는 계약 — DE 책임).
+    // 스케일 규약: price·volume = 8(원천과 동일), amount = price×volume 이라 정확히 16.
+    private BigDecimal tradePrice;  // 체결 가격 (KRW)
+    private BigDecimal tradeVolume; // 체결 수량
+    private BigDecimal tradeAmount; // 체결 금액 = price × volume (MySQL 의 DECIMAL(20,4) 값은 먼지 체결이 0 이라 쓰지 않는다)
     private String askBid;          // ASK(매도) / BID(매수)
     private long upbitTimestamp;    // Upbit 체결 시각 (Unix ms)
     private long sequentialId;      // Upbit 체결 고유 ID
@@ -43,14 +47,14 @@ public class CryptoTradeEvent implements Serializable {
     public String getMarket() { return market; }
     public void setMarket(String market) { this.market = market; }
 
-    public double getTradePrice() { return tradePrice; }
-    public void setTradePrice(double tradePrice) { this.tradePrice = tradePrice; }
+    public BigDecimal getTradePrice() { return tradePrice; }
+    public void setTradePrice(BigDecimal tradePrice) { this.tradePrice = tradePrice; }
 
-    public double getTradeVolume() { return tradeVolume; }
-    public void setTradeVolume(double tradeVolume) { this.tradeVolume = tradeVolume; }
+    public BigDecimal getTradeVolume() { return tradeVolume; }
+    public void setTradeVolume(BigDecimal tradeVolume) { this.tradeVolume = tradeVolume; }
 
-    public double getTradeAmount() { return tradeAmount; }
-    public void setTradeAmount(double tradeAmount) { this.tradeAmount = tradeAmount; }
+    public BigDecimal getTradeAmount() { return tradeAmount; }
+    public void setTradeAmount(BigDecimal tradeAmount) { this.tradeAmount = tradeAmount; }
 
     public String getAskBid() { return askBid; }
     public void setAskBid(String askBid) { this.askBid = askBid; }
